Don't exec if its an db alias.
This is a bad practice, but I was unable to write a good failing case
for this inside the acts_as_eav_model codebase. There is a strange
corner-case that causes all values to be nil if you try to query the DB
on Rails 3.0.x with a where clause that uses a string with a period ie:

FunTimeSchool.where(:foo_bar => "St. Stuff")

The above query will fail without this fix.